test(settings): split page scenarios
